add a looser preview view selection matcher
Rather than requiring markdown to be in htmlToMarkdown format to be
able to create a reference copy, this uses regular expressions to match
a wider range of possible sources.

As a result, this may return multiple matches for limited selections.
This seems preferable than not finding any matches and requiring source
mode for a greater variety.

In order for this to work more ideally, custom processors based on
matchers will need to be added. This will allow for a selection of just
an image to return a result, for example. Additionally, some of the
prefix and suffix code will need to ignore nodes added by obsidian or
plugins that do not reflect the underlying markdown.