add markdown context to quoths and normalize them
Quoths now pull in the context of a range. So if the range starts in the
middle of a bold portion, the opening bold markdown will be included.

This works for:
- bold
- italic
- strikethrough
- code
- blockquote
- list item

Then, the ranges are normalized. Right now that does two things:
1. It removes all blockquoting up to the minimum number of blockquotes.
This means if all lines have at least 2 levels, they will all be reduced
by 2 levels.
2. It removes a list item start if it is only a single line.