fix internal links in embedded mode
Clicking on links goes to the parent markdown-preview-view to handle its
events. Previously, the embed was wrapped in an element with that class
to mimic obsidian style embeds. However, since the quoth embed is not
creating its own preview view, the links went unhandled.

For now, another class is put in instead -- quoth-embedded-view.

In the future, a quoth view may need to be created, it will depend on
what APIs are available.